Add files via upload
This project was a part of my MSc Data Analytics course in which I have used Spark SQL to analyze a dataset around the crime which is taking place in Leicestershire and the City Of London. I have also used Machine Learning Algorithms to predict the survival rate using a Titanic dataset and to compare them based upon a accuracy, precision and recall value.